🔧 rework : ExerciseMuscleGroupController

+++ backend/backend_CalisthenX/src/main/java/org/example/backend_calisthenx/services/ExerciseMuscleGroupService.java
@@ -1,5 +1,6 @@
 package org.example.backend_calisthenx.services;
 
+import org.example.backend_calisthenx.exceptions.ResourceNotFoundException;
 import org.example.backend_calisthenx.models.ExerciseMuscleGroup;
 import org.example.backend_calisthenx.repositories.ExerciseMuscleGroupRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,11 +30,20 @@ public class ExerciseMuscleGroupService {
         return exerciseMuscleGroupRepository.save(exerciseMuscleGroup);
     }
 
-    public ExerciseMuscleGroup updateExerciseMuscleGroup(ExerciseMuscleGroup exerciseMuscleGroup) {
-        return exerciseMuscleGroupRepository.save(exerciseMuscleGroup);
+    public ExerciseMuscleGroup updateExerciseMuscleGroup(Long id,ExerciseMuscleGroup exerciseMuscleGroup) {
+        ExerciseMuscleGroup existing = exerciseMuscleGroupRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("ExerciseMuscleGroup not found with ID: " + id));
+
+        existing.setExercise(exerciseMuscleGroup.getExercise());
+        existing.setMuscleGroup(exerciseMuscleGroup.getMuscleGroup());
+
+        return exerciseMuscleGroupRepository.save(existing);
     }
 
     public void deleteExerciseMuscleGroup(Long id) {
+        if (!exerciseMuscleGroupRepository.existsById(id)) {
+            throw new ResourceNotFoundException("ExerciseMuscleGroup with ID " + id + " not found");
+        }
         exerciseMuscleGroupRepository.deleteById(id);
     }
 }
